+++ library/ui/src/main/java/com/google/android/exoplayer2/ui/DefaultTimeBar.java
@@ -15,6 +15,7 @@
  */
 package com.google.android.exoplayer2.ui;
 
+import android.animation.ValueAnimator;
 import android.content.Context;
 import android.content.res.Resources;
 import android.content.res.TypedArray;
@@ -163,6 +164,9 @@ public class DefaultTimeBar extends View implements TimeBar {
 
   private static final int DEFAULT_INCREMENT_COUNT = 20;
 
+  private static final float SHOWN_SCRUBBER_SCALE = 1.0f;
+  private static final float HIDDEN_SCRUBBER_SCALE = 0.0f;
+
   /**
    * The name of the Android SDK view that most closely resembles this custom view. Used as the
    * class name for accessibility.
@@ -201,6 +205,7 @@ public class DefaultTimeBar extends View implements TimeBar {
   private int lastCoarseScrubXPosition;
   private @MonotonicNonNull Rect lastExclusionRectangle;
 
+  private ValueAnimator scrubberScalingAnimator;
   private float scrubberScale;
   private boolean scrubbing;
   private long scrubPosition;
@@ -326,6 +331,12 @@ public class DefaultTimeBar extends View implements TimeBar {
               / 2;
     }
     scrubberScale = 1.0f;
+    scrubberScalingAnimator = new ValueAnimator();
+    scrubberScalingAnimator.addUpdateListener(
+        animation -> {
+          scrubberScale = (float) animation.getAnimatedValue();
+          invalidate(seekBounds);
+        });
     duration = C.TIME_UNSET;
     keyTimeIncrement = C.TIME_UNSET;
     keyCountIncrement = DEFAULT_INCREMENT_COUNT;
@@ -333,6 +344,44 @@ public class DefaultTimeBar extends View implements TimeBar {
     if (getImportantForAccessibility() == View.IMPORTANT_FOR_ACCESSIBILITY_AUTO) {
       setImportantForAccessibility(View.IMPORTANT_FOR_ACCESSIBILITY_YES);
     }
+  }
+
+  /** Shows the scrubber handle. */
+  public void showScrubber() {
+    showScrubber(/* showAnimationDurationMs= */ 0);
+  }
+
+  /**
+   * Shows the scrubber handle with animation.
+   *
+   * @param showAnimationDurationMs The duration for scrubber showing animation.
+   */
+  public void showScrubber(long showAnimationDurationMs) {
+    if (scrubberScalingAnimator.isStarted()) {
+      scrubberScalingAnimator.cancel();
+    }
+    scrubberScalingAnimator.setFloatValues(scrubberScale, SHOWN_SCRUBBER_SCALE);
+    scrubberScalingAnimator.setDuration(showAnimationDurationMs);
+    scrubberScalingAnimator.start();
+  }
+
+  /** Hides the scrubber handle. */
+  public void hideScrubber() {
+    hideScrubber(/* hideAnimationDurationMs= */ 0);
+  }
+
+  /**
+   * Hides the scrubber handle with animation.
+   *
+   * @param hideAnimationDurationMs The duration for scrubber hiding animation.
+   */
+  public void hideScrubber(long hideAnimationDurationMs) {
+    if (scrubberScalingAnimator.isStarted()) {
+      scrubberScalingAnimator.cancel();
+    }
+    scrubberScalingAnimator.setFloatValues(scrubberScale, HIDDEN_SCRUBBER_SCALE);
+    scrubberScalingAnimator.setDuration(hideAnimationDurationMs);
+    scrubberScalingAnimator.start();
   }
 
   /**
@@ -353,18 +402,6 @@ public class DefaultTimeBar extends View implements TimeBar {
    */
   public void setScrubberColor(@ColorInt int scrubberColor) {
     scrubberPaint.setColor(scrubberColor);
-    invalidate(seekBounds);
-  }
-
-  /**
-   * Sets the scale factor for the scrubber handle. Scrubber enabled size, scrubber disabled size,
-   * scrubber dragged size are scaled by the scale factor. If scrubber drawable is set, the scale
-   * factor isn't applied.
-   *
-   * @param scrubberScale The scale factor for the scrubber handle.
-   */
-  public void setScrubberScale(float scrubberScale) {
-    this.scrubberScale = scrubberScale;
     invalidate(seekBounds);
   }
 
@@ -829,8 +866,8 @@ public class DefaultTimeBar extends View implements TimeBar {
       int playheadRadius = (int) ((scrubberSize * scrubberScale) / 2);
       canvas.drawCircle(playheadX, playheadY, playheadRadius, scrubberPaint);
     } else {
-      int scrubberDrawableWidth = scrubberDrawable.getIntrinsicWidth();
-      int scrubberDrawableHeight = scrubberDrawable.getIntrinsicHeight();
+      int scrubberDrawableWidth = (int) (scrubberDrawable.getIntrinsicWidth() * scrubberScale);
+      int scrubberDrawableHeight = (int) (scrubberDrawable.getIntrinsicHeight() * scrubberScale);
       scrubberDrawable.setBounds(
           playheadX - scrubberDrawableWidth / 2,
           playheadY - scrubberDrawableHeight / 2,